server.c should be done (hopefully) converted some types from utils.h to use uint32_t & int32_t instead of int added a todo list (todo.md, check it regularly, please)

+++ src/dbms/dbmsUtils.c
@@ -50,7 +50,7 @@ int open_keyfile(const int key, const char mode) {
 }
 
 
-int read_value_from_keyfile(const int key_fd, char *value, const size_t size) {
+int read_value_from_keyfile(const int key_fd, char *value, const int size) {
     ssize_t bytes_read;     /* used for error handling of read_line calls */
 
     /* read value */
